google-ADK-quiz-generator/pyproject.toml
[project]
name = "adk_quiz_generator"
version = "0.1.0"
description = "Quiz generator using Google's Agent Development Kit (ADK)"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10"
dependencies = [
    # Core ADK package
    "google-adk>=1.2.0,<2.0.0",
    
    # AI Platform (pinning major version for stability)
    "google-cloud-aiplatform>=1.115.0,<2.0.0",
    
    # Generative AI
    "google-generativeai>=0.8.5,<0.9.0",
    
    # Environment variables
    "python-dotenv>=1.0.1,<2.0.0",
    
    # Tool dependencies
    "requests>=2.32.3,<3.0.0",
    "beautifulsoup4>=4.12.3,<5.0.0",
    "duckduckgo-search>=6.1.8,<7.0.0",
    "prometheus-client>=0.20.0,<0.22.0"
]

[build-system]
requires = ["hatchling>=1.13.0"]
build-backend = "hatchling.build"
